requirements-connectivity.txt
google-genai>=1.20.0